Revert "Feature/front end auth"
