+++ package.json
@@ -3,10 +3,8 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@okta/okta-react": "0.0.9",
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
-    "react-router-dom": "^4.2.2",
     "react-scripts": "1.0.17"
   },
   "scripts": {
@@ -17,6 +15,5 @@
   },
   "devDependencies": {
     "firebase": "^4.8.0"
-  },
-  "proxy": "http://localhost:4000"
+  }
 }
